refactor: move GeckoLib mixins outside of epicfight.mixins.json
