+++ src/main/java/yesman/epicfight/compat/geckolib/mixin/MixinGeoArmorRenderer.java
@@ -1,4 +1,4 @@
-package yesman.epicfight.mixin.geckolib;
+package yesman.epicfight.compat.geckolib.mixin;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
