+++ src/main/resources/epicfight-compat.geckolib.mixins.json
@@ -1,0 +1,12 @@
+{
+  "required": false,
+  "minVersion": "0.8",
+  "compatibilityLevel": "JAVA_21",
+  "package": "yesman.epicfight.compat.geckolib.mixin",
+  "client": [
+    "MixinGeoArmorRenderer"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
